Fix sqlite autoincrement for orgs

+++ api/src/db/models/orgs.py
@@ -1,8 +1,8 @@
 from enum import Enum
 from datetime import datetime
-from src.db.models.__base__ import Base
+from sqlalchemy import Boolean, DateTime, Enum as SqlEnum, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import BigInteger, Boolean, DateTime, Enum as SqlEnum, ForeignKey, String, func
+from src.db.models.__base__ import Base
 
 
 class OrgRole(str, Enum):
@@ -15,7 +15,7 @@ class Org(Base):
     """Represent an organization account in the platform"""
     __tablename__ = 'organizations'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(100), unique=True, nullable=False)
     country: Mapped[str | None] = mapped_column(String(2), nullable=True)
     
@@ -27,10 +27,10 @@ class OrgMember(Base):
     """Represent a member of an organization"""
     __tablename__ = 'organizations_members'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
-    id_org: Mapped[int] = mapped_column(BigInteger, ForeignKey('organizations.id'), nullable=False)
-    id_user: Mapped[int] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=False)
+    id_org: Mapped[int] = mapped_column(Integer, ForeignKey('organizations.id'), nullable=False)
+    id_user: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), nullable=False)
     
     role: Mapped[OrgRole] = mapped_column(SqlEnum(OrgRole, name='org_role'), nullable=False)
     
@@ -43,9 +43,9 @@ class OrgApp(Base):
 
     __tablename__ = 'organizations_apps'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    id_org: Mapped[int] = mapped_column(BigInteger, ForeignKey('organizations.id'), nullable=False)
-    id_app: Mapped[int] = mapped_column(BigInteger, ForeignKey('apps.id'), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id_org: Mapped[int] = mapped_column(Integer, ForeignKey('organizations.id'), nullable=False)
+    id_app: Mapped[int] = mapped_column(Integer, ForeignKey('apps.id'), nullable=False)
 
     active: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default='true')
